Fix Visual Studio compiler warnings

+++ src/drivers/WinAPI/Fl_WinAPI_Screen_Driver.cxx
@@ -486,5 +486,5 @@ int Fl_WinAPI_Screen_Driver::screen_num_unscaled(int x, int y)
 
 
 float Fl_WinAPI_Screen_Driver::base_scale(int numscreen) {
-  return dpi[numscreen][0] / 96.;
+  return float(dpi[numscreen][0] / 96.);
 }
